requirements.txt
Django==4.1.6
djangorestframework==3.14.0
celery==5.2.7
keras
numpy==1.23.4
h5py==3.7.0
opencv-python==4.7.0.68
scikit-learn==1.2.0
Pillow==9.4.0
tensorflow==2.11.0
redis==4.4.0
tensorflow-io
mysqlclient
django-environ